pipeline/requirements.txt
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=0.2.0
